made changes based on the last review
